fix MSVC compile issue

+++ arangod/RocksDBEngine/RocksDBOptionFeature.cpp
@@ -660,7 +660,8 @@ void RocksDBOptionFeature::validateOptions(
 
 void RocksDBOptionFeature::prepare() {
   if (_enforceBlockCacheSizeLimit && _blockCacheSize > 0) {
-    uint64_t shardSize = _blockCacheSize / (1 << _blockCacheShardBits);
+    uint64_t shardSize =
+        _blockCacheSize / (uint64_t(1) << _blockCacheShardBits);
     // if we can't store a data block of the mininmum size in the block cache,
     // we may run into problems when trying to put a large data block into the
     // cache. in this case the block cache may return a Status::Incomplete
